# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /start:
    post:
      operationId: startProcessInstance
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ProcessInstanceInputMessage'
      - in: query
        name: gid
        description: The UUID of the process instance group owned by the party initiating the process
        required: false
        type: string
      - in: query
        name: preceding_gid
        description: The UUID of the preceding process instance group
        required: false
        type: string
      - in: query
        name: collaborationGID
        description: The id of the collaboration group which the process instance group belongs to
        required: false
        type: string
      responses:
        # Response code
        200:
          description: Business Process Instance Started
          schema:
            $ref: '#/definitions/ProcessInstance'
  /start/billofmaterials:
    post:
      operationId: createNegotiationsForBOM
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: bearerToken
        in: header
        type: string
        required: true
      - in: body
        name: billOfMaterial
        required: true
        schema:
          $ref: '#/definitions/BillOfMaterial'
      - in: query
        name: useFrameContract
        description: If this parameter is true and a valid frame contract exists between parties, then an order is started for the product using the details of frame contract
        required: false
        default: false
        type: boolean
      responses:
        200:
          description: Created negotiations for line items for the given party
          schema:
            type: string
        401:
          description: Invalid token. No user was found for the provided token
          schema:
            type: string
        500:
          description: Unexpected error while creating negotiations for line items for the given party
          schema:
            type: string
...
  /collaboration-groups/{id}:
    patch:
      operationId: updateCollaborationGroupName
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      - in: query
        name: groupName
        type: string
        required: true
      responses:
        200:
          description: Updated the collaboration group name successfully
        401:
          description: Invalid token. No user was found for the provided token
        404:
          description: Specified CollaborationGroup does not exist
    delete:
      operationId: deleteCollaborationGroup
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Deleted the collaboration group successfully
        401:
          description: Invalid token. No user was found for the provided token
    get:
      operationId: getCollaborationGroup
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Retrieved the collaboration group successfully
          schema:
            $ref: '#/definitions/CollaborationGroup'
        401:
          description: Invalid token. No user was found for the provided token
        404:
          description: The specified CollaborationGroup does not exist
...
  /collaboration-groups/merge:
    get:
      operationId: mergeCollaborationGroups
      parameters:
        - name: Authorization
          in: header
          type: string
          required: true
        - in: query
          name: bcid
          type: string
          description: Identifier of the base collaboration group
          required: true
        - in: query
          name: cgids
          type: array
          description: List of collaboration group id's to be merged.
          required: true
      responses:
        200:
          description: Merged the collaboration group successfully
          schema:
            $ref: '#/definitions/CollaborationGroup'
        401:
          description: Invalid token. No user was found for the provided token
        400:
          description: Invalid Clause content
...
definitions:
...
  ProcessVariables:
    title: Business Process Variables Object
    type: object
    properties:
      processID:
        type: string
      initiatorID:
        type: string
      responderID:
        type: string
      contentUUID:
        type: string
        description: The UUID of the document in the content variable
      relatedProducts:
        type: array
        items:
          type: string
      relatedProductCategories:
        type: array
        items:
          type: string
      content:
        type: string
      creatorUserID:
        type: string
  ProcessInstanceInputMessage:
    title: Business Process Input Message Object
    type: object
    properties:
      variables:
        $ref: '#/definitions/ProcessVariables'
      processInstanceID:
        type: string
  ProcessInstance:
    title: Business Process Instance
    type: object
    properties:
      processInstanceID:
        type: string
      creationDate:
        type: string
      processID:
        type: string
      status:
        type: string
        enum:
        - STARTED
        - CANCELLED
        - COMPLETED
  ProcessInstanceGroup:
    title: Process Instance Group Object
    type: object
    properties:
      ID:
        type: string
      name:
        type: string
      partyID:
        type: string
      dataChannelId:
        type: string
      processInstanceIDs:
        type: array
        items:
          type: string
      archived:
        type: boolean
      collaborationRole:
        type: string
      firstActivityTime:
        type: string
      lastActivityTime:
        type: string
      precedingProcessInstanceGroup:
        $ref: '#/definitions/ProcessInstanceGroup'
      status:
        type: string
        enum:
        - INPROGRESS
        - CANCELLED
  CollaborationGroup:
    title: Collaboration Group
    type: object
    properties:
      ID:
        type: string
      name:
        type: string
      archived:
        type: boolean
      isArchiveable:
        type: boolean
      status:
        type: string
        enum:
        - INPROGRESS
      isProject:
        type: boolean
      associatedProcessInstanceGroups:
        type: array
        items:
          $ref: '#/definitions/ProcessInstanceGroup'
...
  BillOfMaterial:
    title: Bill of Material Object
    type: object
...